requirements.txt
absl-py==0.11.0
ase==3.20.1
cachetools==4.2.0
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
decorator==4.4.2
et-xmlfile==1.0.1
google-auth==1.24.0
google-auth-oauthlib==0.4.2
googledrivedownloader==0.4
grpcio==1.34.0
h5py==3.1.0
idna==2.10
isodate==0.6.0
jdcal==1.4.1
Jinja2==2.11.2
joblib==1.0.0
kiwisolver==1.3.1
llvmlite==0.35.0
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.3
networkx==2.5
numba==0.52.0
numpy==1.19.4
oauthlib==3.1.0
omegaconf==2.0.5
openpyxl==3.0.5
pandas==1.2.0
Pillow==8.1.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pypng==0.0.20
python-dateutil==2.8.1
python-louvain==0.15
pytz==2020.5
PyYAML==5.3.1
rdflib==5.0.0
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.24.0
scipy==1.6.0
six==1.15.0
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
threadpoolctl==2.1.0
torch==1.7.1+cu110
torch-cluster==1.5.8
torch-geometric==1.6.3
torch-scatter==2.0.5
torch-sparse==0.6.8
torch-spline-conv==1.2.0
torchaudio==0.7.2
torchvision==0.8.2+cu110
tqdm==4.55.1
typing-extensions==3.7.4.3
urllib3==1.26.2
Werkzeug==1.0.1
